working back

+++ src/main/java/com/wash/laundry_app/statistiques/DailyStatsSchedulerService.java
@@ -31,8 +31,8 @@ public class DailyStatsSchedulerService {
     @Transactional
     public void computeStatsForDate(LocalDate date) {
         long nombreCommandes = commandeRepository.countCommandesByDate(date);
-        Double revenuesTotalRaw = commandeRepository.getRevenueByDate(date);
-        BigDecimal revenusTotal = revenuesTotalRaw != null ? BigDecimal.valueOf(revenuesTotalRaw) : null;
+        BigDecimal revenuesTotalRaw = commandeRepository.getRevenueByDate(date);
+        BigDecimal revenusTotal = revenuesTotalRaw != null ? revenuesTotalRaw : null;
         long nombreTapisTraites = commandeRepository.countItemsByDate(date);
 
         StatistiqueJournaliere stat = statRepository.findByDate(date)
